search_playground: abstracting query fields table, search with enter in search box of query mode
